Support Xcode 15.4 builds
## What Changed
- Added a Swift 5.10 manifest path with GRDB 6.29.3, an Xcode-15-compatible FluidAudio revision, and conditional exclusion of Swift-6-only WhisperKit/Nemotron implementations.
- Backported SwiftUI actor isolation and Swift 5.10 concurrency/source syntax across the app while preserving the Xcode 16 behavior.
- Limited Xcode 15 engine choices to Parakeet and Cohere, with persisted unavailable selections falling back safely to Parakeet.
- Added an Xcode 15.4 CI lane, compatibility-specific tests, and source-build documentation.

## Root Intent
The 声 fork must build and run from Xcode 15.4 instead of requiring Swift 6/Xcode 16 dependency manifests and SDK APIs.

## Validation
- Xcode 15.4 debug build
- Xcode 15.4 release build
- 207 Voice Flow compatibility tests
- CLI version and health JSON smoke checks
- Package lock JSON, workflow YAML, and diff whitespace validation

## Scope
Xcode 15.4 supports Parakeet v3/v2/Unified and Cohere. Nemotron and WhisperKit remain available on Xcode 16+ because their current implementations require newer dependency and Core ML APIs.

+++ Sources/MacParakeetCore/Services/Telemetry/CrashReporter.swift
@@ -103,7 +103,7 @@ public final class CrashReporter {
         altStack.withUnsafeMutableBufferPointer { buf in
             var ss = stack_t()
             ss.ss_sp = UnsafeMutableRawPointer(buf.baseAddress!)
-            ss.ss_size = buf.count
+            ss.ss_size = UInt(buf.count)
             ss.ss_flags = 0
             sigaltstack(&ss, nil)
         }
